+++ packages/BiostructureViewer/src/package-test.ts
@@ -9,6 +9,7 @@ import './tests/pdb-tests';
 import './tests/pdb-grid-cell-renderer-tests';
 import './tests/molstar-preview-tests';
 import './tests/viewers';
+import './tests/pdbqt-tests';
 
 // This _package object is for tests only.
 // Call package functions to test code calling package.ts/_package object.